Flytt mye av logikken i BrevPeriodeService til BrevPeriodeUtil
